[dynamo] move bytecode tests in test_misc
[ghstack-poisoned]
